Update d15s01.rpy
The last change I proposed will no longer be valid with this one, the two sentences go together in their contexts and are clearer together to translate. The first sentence has been changed to be more suitable for everyday French: “the way that feels” can be translated as “la façon que je le sent,” but it doesn’t sound French. I emphasized the feeling, so “sensation” seems more appropriate.
The second sentence “you do both” is quite complicated to translate but can be avoided by using the verb “aimer”. Since the subject changes from one character in the first sentence to two in the second, repeating the verb “aimer” does not sound like a repetition in French without losing the original meaning.
The third sentence is more of an expression “i can tell” cannot be translated directly but there is a French equivalent which is “je le vois bien”.